Added ability to present and dismiss without cancel button. This change messed up the layout of activities. Will fix soon.

+++ REActivityViewController/REActivityView.m
@@ -128,7 +128,7 @@ CGFloat const REActivityPageControlHeight		= 10.0f;
 	
 	if (self.activities) {
 		
-		return ceil( self.activities.count / ( self.maxActivitiesPerPage ) );
+		return ceil( (CGFloat)self.activities.count / ( (CGFloat)self.maxActivitiesPerPage ) );
 		
 	} else {
